# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 話術透視鏡 API — Ad Rhetoric X-Ray
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ocr:
    post:
      operationId: processOcr
      tags: [rhetoric]
      summary: Extract text from image via OCR
      description: Accepts a base64-encoded image and returns the recognized text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OcrInput"
      responses:
        "200":
          description: OCR result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OcrResult"
        "400":
          description: Bad request
  /analyze:
    post:
      operationId: analyzeRhetoric
      tags: [rhetoric]
      summary: Analyze ad copy for rhetoric manipulation
      description: Sends text to LLM for rhetoric analysis and saves result to DB
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RhetoricInput"
      responses:
        "200":
          description: Analysis result with annotations, credibility score, and neutral rewrite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Bad request
...
  /records/{id}:
    get:
      operationId: getRecord
      tags: [records]
      summary: Get a single analysis record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Full analysis record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRecord"
        "404":
          description: Record not found
...
components:
  schemas:
...
    OcrInput:
      type: object
      required: [imageBase64, mimeType]
      properties:
        imageBase64:
          type: string
          description: Base64-encoded image data
        mimeType:
          type: string
          description: MIME type of the image (e.g. image/png, image/jpeg)
    OcrResult:
      type: object
      required: [ocrText]
      properties:
        ocrText:
          type: string
          description: Text extracted from the image
    RhetoricAnnotation:
      type: object
      required: [textSpan, category, explanation]
      properties:
        textSpan:
          type: string
          description: The original text fragment that was annotated
        category:
          type: string
          enum:
            - 恐懼訴求
            - 假稀缺
            - 社會認同操控
            - 權威借位
            - 情緒勒索
            - 誇大療效
          description: Rhetoric category
        explanation:
          type: string
          description: Plain-language explanation of what this phrase is doing to the reader
    RhetoricInput:
      type: object
      required: [inputType, textContent, role]
      properties:
        inputType:
          type: string
          enum: [text, image]
          description: Whether the source was typed text or OCR from image
        textContent:
          type: string
          description: The ad copy text to analyze
        role:
          type: string
          enum: [seller, consumer]
          description: >-
            Analysis perspective — seller (賣家：推銷／成交／說服力導向) or
            consumer (消費者：信任／理解／風險導向). Drives an entirely
            different judgment prompt, not just a display label.
    MatchedProduct:
      type: object
      required: [id, name, category, registrationNumber, approvedUses]
      properties:
        id:
          type: number
          description: Reference product ID (see /admin/products)
        name:
          type: string
        category:
          type: string
        registrationNumber:
          type:
            - string
            - "null"
          description: 核准字號
        approvedUses:
          type: string
          description: 核准適應症／功能，用來比對廣告文案是否誇大不實
    RiskLevel:
      type: string
      enum: [低, 中, 高]
      description: Overall risk level of the role-specific judgment
    AnalysisResult:
      type: object
      required: [recordId, originalText, annotations, credibilityScore, neutralRewrite, matchedProducts, role, verdict, coreJudgment, mainStrengths, mainRisks, improvementSuggestions, riskLevel]
      properties:
        recordId:
          type: number
          description: The database record ID for this analysis
        originalText:
          type: string
          description: The original ad copy text
        annotations:
          type: array
          items:
            $ref: "#/components/schemas/RhetoricAnnotation"
        credibilityScore:
          type: number
          description: Overall credibility score from 0 (very manipulative) to 100 (trustworthy)
        neutralRewrite:
          type: string
          description: A neutral, objective rewrite of the ad copy without manipulative language
        matchedProducts:
          type: array
          items:
            $ref: "#/components/schemas/MatchedProduct"
          description: Reference products whose name was found in the analyzed text
        role:
          type: string
          enum: [seller, consumer]
          description: Which perspective this analysis judged the content from（角色）
        verdict:
          type: string
          description: Short one-line verdict for the role's judgment（判定）
        coreJudgment:
          type: string
          description: The core reasoning behind the verdict（核心判斷）
        mainStrengths:
          type: array
          items:
            type: string
          description: Main strengths from the role's perspective（主要優點）
        mainRisks:
          type: array
          items:
            type: string
          description: Main risks from the role's perspective（主要風險）
        improvementSuggestions:
          type: array
          items:
            type: string
          description: Actionable improvement suggestions（改進建議）
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
...
    AnalysisRecord:
      type: object
      required: [id, createdAt, inputType, originalText, annotations, credibilityScore, neutralRewrite, matchedProducts, role, verdict, coreJudgment, mainStrengths, mainRisks, improvementSuggestions, riskLevel]
      properties:
        id:
          type: number
        createdAt:
          type: string
          description: ISO 8601 timestamp
        inputType:
          type: string
          enum: [text, image]
        originalText:
          type: string
        annotations:
          type: array
          items:
            $ref: "#/components/schemas/RhetoricAnnotation"
        credibilityScore:
          type: number
        neutralRewrite:
          type: string
        matchedProducts:
          type: array
          items:
            $ref: "#/components/schemas/MatchedProduct"
          description: Reference products whose name was found in the analyzed text
        role:
          type: string
          enum: [seller, consumer]
          description: Which perspective this analysis judged the content from（角色）
        verdict:
          type: string
          description: Short one-line verdict for the role's judgment（判定）
        coreJudgment:
          type: string
          description: The core reasoning behind the verdict（核心判斷）
        mainStrengths:
          type: array
          items:
            type: string
          description: Main strengths from the role's perspective（主要優點）
        mainRisks:
          type: array
          items:
            type: string
          description: Main risks from the role's perspective（主要風險）
        improvementSuggestions:
          type: array
          items:
            type: string
          description: Actionable improvement suggestions（改進建議）
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
...